+++ config/guru.php
@@ -1,9 +1,0 @@
-<?php
-return [
-    'driver' => env('GURU_DRIVER', 'file'),
-    'events' => [
-//        'an.example.event' => [
-//            EventHandler::class,
-//        ],
-    ],
-];